+++ src/public_api.ts
@@ -2,6 +2,7 @@
  * Public API Surface of ng-hub-ui-avatar
  */
 export * from './lib/avatar-config';
+export * from './lib/avatar-size';
 export * from './lib/avatar.component';
 export * from './lib/avatar.module';
 export * from './lib/avatar.providers';